+++ src/Zicht/Bundle/UrlBundle/Admin/AliasOverviewType.php
@@ -12,17 +12,25 @@ use Symfony\Component\Form\FormInterface;
 use Symfony\Component\Form\FormView;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 use Zicht\Bundle\PageBundle\Entity\Page;
+use Zicht\Bundle\UrlBundle\Url\Provider;
 
-class PageAliasOverviewAdmin extends AbstractType
+class AliasOverviewType extends AbstractType
 {
+    /** @var Provider */
+    protected $provider;
+
     /** @var RegistryInterface */
     protected $doctrine;
 
     /**
-     * @{inheritDoc}
+     * AliasOverviewType constructor.
+     *
+     * @param Provider $provider
+     * @param RegistryInterface $doctrine
      */
-    function __construct(RegistryInterface $doctrine)
+    function __construct(Provider $provider, RegistryInterface $doctrine)
     {
+        $this->provider = $provider;
         $this->doctrine = $doctrine;
     }
 
@@ -32,10 +40,10 @@ class PageAliasOverviewAdmin extends AbstractType
     public function configureOptions(OptionsResolver $resolver)
     {
         parent::configureOptions($resolver);
-        $resolver->setRequired('page');
+        $resolver->setRequired('record');
         $resolver->setDefaults([
             'translation_domain' => 'admin',
-            'label' => 'admin.page_alias_overview_admin.label',
+            'label' => 'admin.alias_overview_admin.label',
             'required' => false,
             'virtual' => true,
         ]);
@@ -47,8 +55,8 @@ class PageAliasOverviewAdmin extends AbstractType
     public function buildView(FormView $view, FormInterface $form, array $options)
     {
         parent::buildView($view, $form, $options);
-        $view->vars['page'] = $options['page'];
-        $view->vars['url_aliases'] = $this->getUrlAliases($options['page']);
+        $view->vars['record'] = $options['record'];
+        $view->vars['url_aliases'] = $this->getUrlAliases($options['record']);
 
 //        dump($view->vars);
 //        die(sprintf('%s:%d', __FILE__, __LINE__));
@@ -60,7 +68,8 @@ class PageAliasOverviewAdmin extends AbstractType
      */
     protected function getUrlAliases(Page $page)
     {
-        $internalUrl = sprintf('/%s/page/%d', $page->getLanguage(), $page->getId());
+        $internalUrl = $this->provider->url($page);
+//        $internalUrl = sprintf('/%s/page/%d', $page->getLanguage(), $page->getId());
         return $this->doctrine->getRepository('ZichtUrlBundle:UrlAlias')->findAllByInternalUrl($internalUrl);
     }
 
@@ -69,6 +78,6 @@ class PageAliasOverviewAdmin extends AbstractType
      */
     public function getName()
     {
-        return 'page_alias_overview_admin';
+        return 'alias_overview_type';
     }
 }