+++ src/Zicht/Bundle/UrlBundle/Aliasing/Aliasing.php
@@ -180,13 +180,6 @@ class Aliasing
         $ret = false;
         /** @var $alias UrlAlias */
 
-        if (!in_array($conflictingInternalUrlStrategy, [self::STRATEGY_IGNORE, self::STRATEGY_MOVE_PREVIOUS_TO_NEW])) {
-            throw new \InvalidArgumentException("Invalid \$conflictingInternalUrlStrategy '$conflictingInternalUrlStrategy'");
-        }
-        if (!in_array($conflictingPublicUrlStrategy, [self::STRATEGY_KEEP, self::STRATEGY_OVERWRITE, self::STRATEGY_SUFFIX])) {
-            throw new \InvalidArgumentException("Invalid \$conflictingPublicUrlStrategy '$conflictingPublicUrlStrategy'");
-        }
-
         // if the internal url is currently already aliased
         $alias = $this->hasPublicAlias($internalUrl, true);
 
@@ -202,12 +195,11 @@ class Aliasing
                 case self::STRATEGY_IGNORE:
                     // Alias already exist, but the strategy is to ignore changes
                     return $ret;
-                    break;
                 default:
                     // case is handled in the 'if' guard at top of the function
+                    break;
             }
         }
-
 
         if ($alias = $this->hasInternalAlias($publicUrl, true)) {