+++ CHANGELOG.md
@@ -9,4 +9,4 @@
 - Reworked the rewriting internals for alias processing by eliminating a lot of duplicated code.
 - Add tests
 # 2.10.0 #
-- Adds PageAliasOverviewAdmin which shows all public urls of a page in the CMS.
+- Adds AliasOverviewType which shows all public urls of a page in the CMS.